Timer, <script> and java, no countdown, but working on it
If you notice any strange code, it was just me working on the timer. Feel free to edit or remove it, just contributing to the cause that needs to be worked on a bit. No countdown, but I'm working on that as well. Hopefully up by tomorrow or in a few days. 

- Max S.